requirements.txt
yt-dlp>=2024.0.0
youtube-transcript-api>=0.6.0
google-genai>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
weasyprint>=60.0
markdown>=3.5
pymdown-extensions>=10.0
pyyaml>=6.0
tqdm>=4.60
tiktoken>=0.5.0
pytest>=7.0
